requirements.txt
flask
openai
python-dotenv
spacy
spacy-langdetect